refactor: extract KeychainSessionDataProtector.swift from SessionDataProtector.swift
Splits the `KeychainSessionDataProtector` struct (~60 lines) out of
SessionDataProtector.swift into its own file. SessionDataProtector.swift
retains the protocol, the PlaintextSessionDataProtector fallback, and
the SessionDataProtectorFactory selector.

Byte-identical move of lines 23-82. Both files import CryptoKit +
Foundation + Security.

SessionDataProtector.swift: 95 → 33 lines. Tests: 667.
